+++ src/main/java/erp/greenagro/greenagro_erp_backend/strategy/product_update/NormalUpdateStrategy.java
@@ -1,17 +1,17 @@
-package erp.greenagro.greenagro_erp_backend.strategy;
+package erp.greenagro.greenagro_erp_backend.strategy.product_update;
 
 import erp.greenagro.greenagro_erp_backend.dto.product.UpdateProductRequest;
 import erp.greenagro.greenagro_erp_backend.model.entity.Product;
 import erp.greenagro.greenagro_erp_backend.model.entity.ProductGroup;
 import org.springframework.stereotype.Component;
 
-@Component("NORMAL")
+@Component("NORMAL_UPDATE")
 public class NormalUpdateStrategy implements ProductDetailUpdateStrategy{
 
     // ? -> 일반 으로 그룹변경 할때
 
     @Override
-    public void updateDetails(Product product, ProductGroup originGroup, ProductGroup updateGroup, UpdateProductRequest request) {
+    public void updateDetail(Product product, ProductGroup originGroup, ProductGroup updateGroup, UpdateProductRequest request) {
 
         System.out.println("전략: 일반 선택 됨");
